% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_data_set}
\alias{random_data_set}
\title{A simple uniform random data set}
\usage{
random_data_set(n_rows = 1000, n_cols = 5)
}
\arguments{
\item{n_rows}{How many rows}

\item{n_cols}{How many columns}
}
\value{
A data frame with numeric types and simple column names
  (COL_1, COL_2 etc)
}
\description{
A simple uniform random data set
}
\examples{
random_data_set(10, 3)
}
